Add ubuntu 15.10 and delete 15.04

+++ scripts/ubuntu/cleanup.sh
@@ -8,8 +8,11 @@ echo "cleaning up dhcp leases"
 rm /var/lib/dhcp/*
 
 echo "cleaning up udev rules"
-rm /etc/udev/rules.d/70-persistent-net.rules
+if [ -f /etc/udev/rules.d/70-persistent-net.rules ]; then
+  rm /etc/udev/rules.d/70-persistent-net.rules
+fi
 mkdir /etc/udev/rules.d/70-persistent-net.rules
 rm -rf /dev/.udev/
-rm /lib/udev/rules.d/75-persistent-net-generator.rules
-
+if [ -f /lib/udev/rules.d/75-persistent-net-generator.rules ]; then
+  rm /lib/udev/rules.d/75-persistent-net-generator.rules
+fi
